+++ src/main/java/com/strangegrotto/wealthdraft/tax/RegularIncomeTaxCalculator.java
@@ -1,7 +1,10 @@
 package com.strangegrotto.wealthdraft.tax;
 
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.strangegrotto.wealthdraft.govconstants.GovConstantsForYear;
+import com.strangegrotto.wealthdraft.scenarios.ImmutableIncomeStreams;
 import com.strangegrotto.wealthdraft.scenarios.IncomeStreams;
 import com.strangegrotto.wealthdraft.scenarios.Scenario;
 import org.slf4j.Logger;
@@ -25,67 +28,82 @@ public class RegularIncomeTaxCalculator {
         // We don't apply the standard deduction yet though; that gets added later
         Deductions deductions = DeductionsCalculator.calculateAllowedDeductions(scenario, govConstants);
         log.debug("Deductions: {}", deductions);
+
         long retirementDeductions = deductions.getTrad401kDeduction() +
                 deductions.getTradIraDeduction();
-        log.debug("Trad 401k & IRA deductions: {}", retirementDeductions);
-        log.debug("Standard Deduction: {}", deductions.getStandardDeduction());
-
         IncomeStreams taxableIncome = DeductionsCalculator.applyDeduction(scenario.getIncomeStreams(), retirementDeductions);
-        log.debug("Reg Fed Taxable Income (includes FEI, lacks standard deduction): {}", taxableIncome);
+        log.debug("Reg Fed Taxable Income (includes FEI, std deduc NOT applied): {}", taxableIncome);
+
+        long excludedFEI = Math.min(
+                govConstants.getForeignIncomeConstants().getForeignEarnedIncomeExemption(),
+                (long)((double)taxableIncome.getEarnedIncome() * scenario.getFractionForeignEarnedIncome())
+        );
+        log.debug("Excluded FEI: {}", excludedFEI);
+        taxableIncome = applyStdDeductionGivenFEI(
+                taxableIncome,
+                excludedFEI,
+                govConstants.getStandardDeduction()
+        );
+        log.debug("Reg Fed Taxable Income (includes FEI, std deduc IS applied): {}", taxableIncome);
 
         long earnedIncome = taxableIncome.getEarnedIncome();
-        long otherUnearnedIncome = taxableIncome.getOtherUnearnedIncome();
-        long stcg = taxableIncome.getShortTermCapGains();
-        long ltcg = taxableIncome.getLongTermCapGains();
+        long nonPrefUnearnedIncome = taxableIncome.getNonPreferentialUnearnedIncome();
+        long prefUnearnedIncome = taxableIncome.getPreferentialUnearnedIncome();
 
         ImmutableMap.Builder<Tax, Double> totalTaxes = ImmutableMap.builder();
 
         // Earned income tax
-        long nonPreferentialIncome = earnedIncome + otherUnearnedIncome + stcg;
-        long excludedFEI = Math.min(
-                govConstants.getForeignIncomeConstants().getForeignEarnedIncomeExemption(),
-                (long)((double)earnedIncome * scenario.getFractionForeignEarnedIncome())
-        );
-        long nonPreferentialNonExcludedFEIIncome = nonPreferentialIncome - excludedFEI;
-        log.debug("Non preferential income: {}", nonPreferentialIncome);
-        log.debug("Excluded FEI: {}", excludedFEI);
-        log.debug("Non preferential non excluded income: {}", nonPreferentialNonExcludedFEIIncome);
-
-        // We apply the standard deduction separately here, rather than at the start, because the standard deduction
-        //  gets applied AFTER the FEI exclusion. If we applied the standard deduction first, we'd potentially be
-        //  "wasting" it on FEI that was going to get excluded anyways
-        long stdDeductionUsedOnNonPrefNonFEIEIncome = Math.min(
-                deductions.getStandardDeduction(),
-                nonPreferentialNonExcludedFEIIncome
-        );
-        long nonPrefNonFEIETaxableIncome = Math.max(
-                0,
-                nonPreferentialNonExcludedFEIIncome - stdDeductionUsedOnNonPrefNonFEIEIncome
-        );
-        long remainingStdDeduction = deductions.getStandardDeduction() - stdDeductionUsedOnNonPrefNonFEIEIncome;
-        log.debug("Non preferential nonexcluded income less std deduction: {}", nonPrefNonFEIETaxableIncome);
+        long nonPrefIncome = earnedIncome + nonPrefUnearnedIncome;
+        long prefIncome = prefUnearnedIncome; // Is there even pref, earned income??
 
         ProgressiveTaxCalculator taxCalculator = new ProgressiveTaxCalculator(govConstants.getFederalIncomeTaxBrackets());
-        long nonPreferentialTaxableIncome = excludedFEI + nonPrefNonFEIETaxableIncome;
         double excludedFEITax = taxCalculator.calculateTax(excludedFEI);
-        double nonPrefIncomeWithExcludedFEITax = taxCalculator.calculateTax(nonPreferentialTaxableIncome);
-        double nonPrefTax = nonPrefIncomeWithExcludedFEITax - excludedFEITax;
-        totalTaxes.put(Tax.NON_PREFERENTIAL_INCOME, nonPrefTax);
+        double nonPrefIncomeTax = taxCalculator.calculateTax(nonPrefIncome);
+        double nonPrefTaxWithFEIE = nonPrefIncomeTax - excludedFEITax;
+        totalTaxes.put(Tax.FED_NON_PREF_INCOME, nonPrefTaxWithFEIE);
 
         // Preferential capital gains: these are "stacked" on top of other income, so unfortunately they don't start at the absolute
         //  lowest rate
         ProgressiveTaxCalculator fedLtcgTaxCalculator = new ProgressiveTaxCalculator(govConstants.getFederalLtcgBrackets());
-        long preferentialTaxableIncome = Math.max(
-                0,
-                ltcg - remainingStdDeduction
-        );
-        log.debug("Preferential income less remaining std deduction: {}", preferentialTaxableIncome);
-        double ltcgPlusIncomeUnderneathTax = fedLtcgTaxCalculator.calculateTax(nonPreferentialTaxableIncome + preferentialTaxableIncome);
-        double incomeUnderneathLtcgTax = fedLtcgTaxCalculator.calculateTax(nonPreferentialTaxableIncome);
-        double ltcgTax = ltcgPlusIncomeUnderneathTax - incomeUnderneathLtcgTax;
-        totalTaxes.put(Tax.PREFERENTIAL_INCOME, ltcgTax);
+        double prefPlusNonPrefLtcgTax = fedLtcgTaxCalculator.calculateTax(prefIncome + nonPrefIncome);
+        double nonPrefLtcgTax = fedLtcgTaxCalculator.calculateTax(nonPrefIncome);
+        double prefTax = prefPlusNonPrefLtcgTax - nonPrefLtcgTax;
+        totalTaxes.put(Tax.FED_PREF_INCOME, prefTax);
 
         return totalTaxes.build();
+    }
 
+    /**
+     * We can't just use {@link DeductionsCalculator#applyDeduction} as-is with the standard deduction because
+     * then it will potentially reduce dollars which should have been excluded under foreign-earned income.
+     * @param income Income streams to reduce
+     * @param excludedForeignEarnedIncome The amount of foreign-earned income that can actually be excluded
+     * @param stdDeduction Standard deduction
+     * @return Income streams with standard deduction applied (given the FEIE)
+     */
+    @VisibleForTesting
+    static IncomeStreams applyStdDeductionGivenFEI(
+            IncomeStreams income,
+            // Of the foreign-earned income, how much is actually excluded based on FEIE and time spent abroad
+            long excludedForeignEarnedIncome,
+            long stdDeduction) {
+        Preconditions.checkArgument(
+                excludedForeignEarnedIncome <= income.getEarnedIncome(),
+                "Excluded foreign-earned income cannot be greater than earned income!");
+
+        // To apply the standard deduction, we need to subtract out the FEI and then add it back
+        //  after the deduction is applied (so the std deduction doesn't erroneously reduce any FEI excluded
+        //  dollars)
+        IncomeStreams incomeLessFEI = ImmutableIncomeStreams.builder()
+                .earnedIncome(income.getEarnedIncome() - excludedForeignEarnedIncome)
+                .nonPreferentialUnearnedIncome(income.getNonPreferentialUnearnedIncome())
+                .preferentialUnearnedIncome(income.getPreferentialUnearnedIncome())
+                .build();
+        IncomeStreams afterDeduction = DeductionsCalculator.applyDeduction(incomeLessFEI, stdDeduction);
+        return ImmutableIncomeStreams.builder()
+                .earnedIncome(afterDeduction.getEarnedIncome() + excludedForeignEarnedIncome)
+                .nonPreferentialUnearnedIncome(afterDeduction.getNonPreferentialUnearnedIncome())
+                .preferentialUnearnedIncome(afterDeduction.getPreferentialUnearnedIncome())
+                .build();
     }
 }